rework conversation state
